convert tests to test cli command
